requirements.txt
astroid==1.6.2
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pylint==1.8.3
python-telegram-bot==10.0.1
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
